[FLINK-6841][table]Using TableSourceTable for both Stream and Batch
