+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/plan/schema/TableSourceTable.scala
@@ -18,9 +18,11 @@
 
 package org.apache.flink.table.plan.schema
 
+import org.apache.calcite.rel.`type`.{RelDataType, RelDataTypeFactory}
 import org.apache.flink.table.api.TableEnvironment
+import org.apache.flink.table.calcite.FlinkTypeFactory
 import org.apache.flink.table.plan.stats.FlinkStatistic
-import org.apache.flink.table.sources.TableSource
+import org.apache.flink.table.sources.{DefinedProctimeAttribute, DefinedRowtimeAttribute, TableSource}
 
 /** Table which defines an external table via a [[TableSource]] */
 class TableSourceTable[T](
@@ -30,4 +32,37 @@ class TableSourceTable[T](
     typeInfo = tableSource.getReturnType,
     fieldIndexes = TableEnvironment.getFieldIndices(tableSource),
     fieldNames = TableEnvironment.getFieldNames(tableSource),
-    statistic)
+    statistic){
+
+    override def getRowType(typeFactory: RelDataTypeFactory): RelDataType = {
+        val flinkTypeFactory = typeFactory.asInstanceOf[FlinkTypeFactory]
+
+        val fieldNames = TableEnvironment.getFieldNames(tableSource).toList
+        val fieldTypes = TableEnvironment.getFieldTypes(tableSource.getReturnType).toList
+
+        val fieldCnt = fieldNames.length
+
+        val rowtime = tableSource match {
+            case timeSource: DefinedRowtimeAttribute if timeSource.getRowtimeAttribute != null =>
+                val rowtimeAttribute = timeSource.getRowtimeAttribute
+                Some((fieldCnt, rowtimeAttribute))
+            case _ =>
+                None
+        }
+
+        val proctime = tableSource match {
+            case timeSource: DefinedProctimeAttribute if timeSource.getProctimeAttribute != null =>
+                val proctimeAttribute = timeSource.getProctimeAttribute
+                Some((fieldCnt + (if (rowtime.isDefined) 1 else 0), proctimeAttribute))
+            case _ =>
+                None
+        }
+
+        flinkTypeFactory.buildLogicalRowType(
+            fieldNames,
+            fieldTypes,
+            rowtime,
+            proctime)
+
+    }
+}
